requirements.txt
CherryPy==3.2.2
Jinja2==2.6
Unidecode==0.04.9
beautifulsoup4==4.0.1
certifi==0.0.8
chardet2==2.0.1
distribute==0.6.19
ipdb==0.7dev
ipython==0.12
pymongo3==1.9b1
requests==0.10.8
wsgiref==0.1.2
